Fix insane bug where some profiles cause client hang
Some code was converting the customs object to an array with millions of elements, which caused iteration and ipc transport (json serialize) to hang the client for many minutes at a time.

This is probably because incorrect lodash functions were used on the customs. Lodash would convert it to array and then it would get stored and turn up as array every time, corrupting storage and ensuring client denial of service.

This fixes that bug by avoiding using incorrect lodash functions that convert things to arrays with millions of elements.

It also retroactively fixes profiles that were stored with array customs, so the client won't crash.