fix: Duplicate event listeners
Cache startup was being run more than once, it seems more correct to
have that function called when a connection is established (as in
chat/Chat.vue) rather than here, otherwise relogging will break them.

Also fixes the handling of duplicate event listeners in the event bus,
although this should not strictly be necessary and the solution in place
is rather fragile. Comparing two functions like this will never have
equality unless converted to a primitive.