Grouping and seperators that obey grouping (Need to do some optimising)
So, this solution is a little fucked and honestly I'm not 100% sure it's optimal. The RAM cost is probably not great since it stores the entirety of the message before it inside of every message now (unless it is undefined) but this works for now and there's plenty of time to poke at it.