Few more changes to Modern Chat
Defaults grouping to true.
Implements time on hover for Modern Chat UI.
Makes some opinionated changes to Modern Chat UI.
Fixes(?) a bug in common where noDate in formatTime doesn't remove date, I'm not 100% certain this is a change that is ideal or if I'm better making a new function so gonna investigate more to be sure.